backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ultralytics==8.1.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
